pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "the library determining the difference between two data structures"
authors = ["slovohot <i@slovohot.ru>"]
readme = "README.md"
packages = [{include = "gendiff"}]


classifiers = [
    "Programming Language :: Python :: 3.11",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]


[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.0"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pyyaml = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
